Add a trait with most of the implementation
This makes it easier to add Prophecy support when needing to extend a
testcase providing other features (many frameworks are shipping their
own base test cases).
